requirements.txt
black==22.3.0
google==3.0.0
google_api_python_client==2.77.0
nextcord==2.3.2
python-dotenv==0.21.1